+++ tests/testthat/test-content-tools.R
@@ -29,7 +29,7 @@ test_that("invoke_tool returns a ContentToolResult", {
   expect_s3_class(res, "ellmer::ContentToolResult")
   expect_s3_class(res@error, "condition")
   expect_true(tool_errored(res))
-  expect_equal(tool_error_string(res), "unused argument (x = 1)")
+  expect_match(tool_error_string(res), "unused argument", ignore.case = TRUE)
   expect_equal(res@value, NULL)
   expect_equal(res@extra, list())
   expect_s3_class(res@request, "ellmer::ContentToolRequest")
@@ -110,7 +110,7 @@ test_that("invoke_tool_async returns a ContentToolResult", {
   expect_s3_class(res, "ellmer::ContentToolResult")
   expect_s3_class(res@error, "condition")
   expect_true(tool_errored(res))
-  expect_equal(tool_error_string(res), "unused argument (x = 1)")
+  expect_match(tool_error_string(res), "unused argument", ignore.case = TRUE)
   expect_equal(res@value, NULL)
   expect_equal(res@extra, list())
   expect_s3_class(res@request, "ellmer::ContentToolRequest")
@@ -172,6 +172,48 @@ test_that("invoke_tools_async() echoes tool requests and results", {
 
   expect_silent(sync(invoke_tools_async(turn)))
   expect_snapshot(. <- sync(invoke_tools_async(turn, echo = "output")))
+})
+
+test_that("invoke_tools() converts to R data structures", {
+  out <- NULL
+  tool <- tool(
+    function(...) out <<- list(...),
+    "A tool",
+    x = type_array(items = type_number()),
+    y = type_array(items = type_string())
+  )
+
+  res <- invoke_tool(
+    ContentToolRequest(
+      id = "x",
+      name = "my_tool",
+      arguments = list(x = list(1, 2), y = list()),
+      tool = tool
+    )
+  )
+  expect_equal(out$x, c(1, 2))
+  expect_equal(out$y, character())
+})
+
+test_that("invoke_tools_async() converts to R data structures", {
+  out <- NULL
+  tool <- tool(
+    function(...) out <<- list(...),
+    "A tool",
+    x = type_array(items = type_number()),
+    y = type_array(items = type_string())
+  )
+
+  res <- sync(invoke_tool_async(
+    ContentToolRequest(
+      id = "x",
+      name = "my_tool",
+      arguments = list(x = list(1, 2), y = list()),
+      tool = tool
+    )
+  ))
+  expect_equal(out$x, c(1, 2))
+  expect_equal(out$y, character())
 })
 
 test_that("tool error warnings", {